// src/models/permission.ts
export interface PermissionData {
  no_thing?: number;
  EzTradeChargeRate?: number;
  EzTrade?: number;
  EzTransfer?: number;
  EzAdvance?: number;
  EzMargin?: number;
  EzMortgage?: number;
  EzOddlot?: number;
  EzMarginPro?: number;
  EzFuture?: number;
  EzTvdt?: number;
  vTblid?: number;
  EzSaving?: number;
  EzMarginPlus?: number;
  Adomestic?: number;
  vFeeUP?: number;
  vFeeUP_CCQ?: number;
  vFeeLISTED_CP?: number;
  vFeeHSX_CP?: number;
  vFeeRate_TP?: number;
  vFeeLISTED_ETF?: number;
  vFeeLISTED_CCQ?: number;
  vFeeHSX_CCQ?: number;
  vFeeHSX_CQ?: number;
  vFeeHSX_ETF?: number;
  vFeeLISTED_CQ?: number;
}
export interface SessionData{
  LoginName?: string
  Password?: any
  TradingPassword?: any
  Source?: any
  IpClient?: string
  IpServer?: string
  SessionToken?: string
  Referer?: string
  UserAgent?: string
  Browser?: any
  IsMobile?: any
  BrowserName?: any
  BrowserVers?: any
  ClientCode?: string
  ClientName?: string
  SessionNo?: string
  ErrorCode?: number
  Message?: string
  CheckSession?: number
  LoginStatus?: number
  RsaToken?: number
  CheckTradingPassword?: number
  RsaResult?: number
  Jwt?: any
  IsCheckSessionSuccess?: boolean
  IsFirstLogin?: boolean
  IsRsaTokenTradingPassword?: boolean
  RsaMessage?: string
  CheckPass2?: number
  TradingPass1By1?: number
  OldPass?: any
  NewPass?: any
  OldTradePass?: any
  NewTradePass?: any
  IntResult?: number
  Areason?: number
  Language?: any
  Activity?: any
  fromDate?: any
  toDate?: any
  Arsatokenupdtype?: any
  Arsatokenupddesc?: any
  Aupduser?: any
  Aupdtbtntype?: any
  Brkid?: any
  Asupportpass?: number
  Permit?: any
  OTPVerifed?: number
  SMS?: string
  ATRANSACTION?: number
  smsOtp?: any
  Resend?: number
  CountOtpFalse?: number
  StattusOtp?: number
  TypeOtp?: any
  TimeExpire?: any
  TimeReSend?: string
  TimeServer?: string
  RetryTimesMax?: number
  SendTimes?: number
  Email?: string
}